Move DB metrics export function to a separate file
